+++ public/dialog_font_size.js
@@ -1,3 +1,4 @@
+import {ChangeEditorOptionEvent} from '/change_editor_option_event.js';
 import {DialogPrompt} from '/dialog_prompt.js';
 import {Res} from '/res.js';
 
@@ -9,13 +10,15 @@ export class DialogFontSize extends DialogPrompt {
     }
 
     onChange(event) {
-        this.core.setFontSize(event.target.value + 'px');
+        document.dispatchEvent(
+            new ChangeEditorOptionEvent('fontSize', event.target.value + 'px')
+        );
     }
 
     open() {
         const res = new Res();
         super.open(
-            res.titles.font_size, null, res.placeholders.font_size,
+            res.descriptions.font_size, null, res.placeholders.font_size,
             this.initialValue, 'number', 8, 72
         );
     }
@@ -27,7 +30,9 @@ export class DialogFontSize extends DialogPrompt {
 
     reset() {
         super.reset();
-        this.core.setFontSize(this.initialValue + 'px');
+        document.dispatchEvent(
+            new ChangeEditorOptionEvent('fontSize', this.initialValue + 'px')
+        );
         this.core.saveRendererOptions();
     }
 }