+++ public/res/CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## Version 2.0.30200415 (dev)
+- Fixed a bug that Theme dialog doesn't open.
 - Updated ace-builds to version 1.4.10.
 
 ## Version 2.0.30200414 (dev)