+++ public/res/CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## Version 2.0.30200416 (dev)
+- Cleaned up all title strings.
+
 ## Version 2.0.30200415 (dev)
 - Fixed the menu name to refer to res.js.
 - Fixed the option name same as Ace Editor.